requirements.txt
python-dotenv
openai
pandas
ccxt